feature -> develop | Fixed migration

+++ Database/Migrations/2021_08_17_132505_vh_cms_content_form_fields_add_vh_cms_form_group_index.php
@@ -30,8 +30,10 @@ class VhCmsContentFormFieldsAddVhCmsFormGroupIndex extends Migration
     */
     public function down()
     {
-        Schema::table('vh_cms_content_form_fields', function (Blueprint $table) {
-            $table->dropColumn('vh_cms_form_group_index');
-        });
+        if (Schema::hasColumn('vh_cms_content_form_fields', 'vh_cms_form_group_index')) {
+            Schema::table('vh_cms_content_form_fields', function (Blueprint $table) {
+                $table->dropColumn('vh_cms_form_group_index');
+            });
+        }
     }
 }
